+++ Assets/Scripts/UI/UIManager.cs
@@ -22,10 +22,10 @@ public class UIManager : MonoBehaviour
         _levelText.text = FixedValue.LEVEL_NAME[level];
         _waterdropGage.fillAmount = (float)waterdrop / FixedValue.MAX_GAGE[level];
     }
-    public void SetEnvStatUI(int temper, float water)
+    public void SetEnvStatUI(float temper, float water)
     {
         _waterGage.fillAmount = water;
-        _temperGage.fillAmount = (float)temper / FixedValue.MAX_TEMPER;
+        _temperGage.fillAmount = temper / (float)FixedValue.MAX_TEMPER;
     }
     public void SetMoneyUI(int amt)
     {
